Introduce Mocha to test test/utils.js
Also enable colors in GitHub Actions for nicer logs.

Fixes https://github.com/mdn/browser-compat-data/issues/6855.
