+++ .github/workflows/node.js.yml
@@ -26,3 +26,5 @@ jobs:
         node-version: ${{ matrix.node-version }}
     - run: npm install
     - run: npm test
+      env:
+        FORCE_COLOR: 3
