+++ package.json
@@ -35,19 +35,21 @@
     "chalk": "~4.1.0",
     "compare-versions": "~3.6.0",
     "mdn-confluence": "~2.2.0",
+    "mocha": "~8.2.0",
     "ora": "~5.1.0",
     "prettier": "~2.1.2",
     "yargs": "~16.1.0"
   },
   "scripts": {
     "confluence": "node ./node_modules/mdn-confluence/main/generate.es6.js --output-dir=. --bcd-module=./index.js",
+    "mocha": "mocha \"test/**.test.js\"",
     "lint": "node test/lint",
     "fix": "node scripts/fix",
     "mirror": "node scripts/mirror",
     "stats": "node scripts/statistics",
     "release-notes": "node scripts/release-notes",
     "show-errors": "npm test 1> /dev/null",
-    "test": "npm run lint",
+    "test": "npm run mocha && npm run lint",
     "traverse": "node scripts/traverse"
   }
 }
